requirements/development.txt
--requirement=common.txt

django-debug-toolbar==1.9.1
ipython==2.1.0

flake8==2.0
pep8==1.4.6
